Rozbiłem app.py na dwa pliki dla łatwiejszego testowania
